Added void* extra to sokol_main
The reason for this is that in my Android build I need to some stuff with the NativeActivity object as soon as possible, even before Sokol starts doing anything. But for now I am relying on Sokol_App to manage the application creation, so the "best" way to get access to the ANativeActivity is like this. I just pass a NULL on all other implementations tho, so it's only used for android so far... But I can imagine it being usefull in other contexts too.